cleanup: move ast/consts to fuzionconstants
